Modificando calculadora

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -5,16 +5,16 @@ import java.util.Scanner;
 public class Calculadora {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        int a, b;
+        double a, b;
         System.out.println("Digite o primeiro valor: ");
         a= scan.nextInt();
          System.out.println("Digite o Segundo valor: ");
         b= scan.nextInt();
 
-        int somar = somar(a , b);
-        int subtrair = subtrair(a , b);
-        int multiplicar = multiplicar(a , b);
-        int dividir = dividir(a , b);
+        double somar = somar(a , b);
+        double subtrair = subtrair(a , b);
+        double multiplicar = multiplicar(a , b);
+        double dividir = dividir(a , b);
 
         System.out.println("Somar: "+somar);
         System.out.println("Subtrair: "+subtrair);
@@ -28,16 +28,16 @@ public class Calculadora {
 
     }
 
-    public static int somar(int a, int b){
+    public static double somar(double a, double b){
         return a+b;
     }
-    public static int subtrair(int a, int b){
+    public static double subtrair(double a, double b){
         return a-b;
     }
-    public static int multiplicar(int a, int b) {
+    public static double multiplicar(double a, double b) {
         return a*b;
     }
-    public static int dividir(int a, int b){
+    public static double dividir(double a, double b){
         return a/b;
     }
